feat(algo): the medial axis — the audit's last absent row closes
`ogeom_algo::medial_axis`: the locus of centres of maximal inscribed
circles, for convex planar polygonal faces, exactly. The shrinking-
polygon construction — every edge inward at unit speed, every vertex
riding its angular bisector, each event retiring an edge and starting a
branch — with convexity as the honesty condition, because convexity is
precisely what excludes the split events the construction does not have.

Held to closed forms older than the field: a 20×10 rectangle's axis is
its four corner diagonals and roof line, total length 20√2 + 10, meets at
half height, deepest clearance exactly the half height; a 3–4–5
triangle's branches meet at the incenter and the deepest clearance is the
inradius (a + b − c)/2 = 1. Holes, reflex corners and curved boundaries
are refused by name — split events and parabolic bisectors each change
the mathematics, and a wrong axis gouges tool paths quietly.

With this, the parity ledger holds no `absent` verdict at all: 69
covered, 13 partial with stated restrictions, 9 divergent with written
reasoning, 5 n/a, 1 unreviewed. Every gap that remains is a named
restriction on a working capability, not a missing one.

Gate green end to end.

+++ crates/ogeom-io/src/native.rs
@@ -884,10 +884,10 @@ pub fn read_document(text: &str) -> OgeomResult<ogeom_doc::Document> {
                 } else {
                     None
                 };
-                let lines = cursor.count()? as usize;
+                let lines = cursor.count()?;
                 let mut polylines = Vec::with_capacity(lines);
                 for _ in 0..lines {
-                    let count = cursor.count()? as usize;
+                    let count = cursor.count()?;
                     let mut line = Vec::with_capacity(count);
                     for _ in 0..count {
                         line.push(cursor.point()?);
@@ -895,9 +895,9 @@ pub fn read_document(text: &str) -> OgeomResult<ogeom_doc::Document> {
                     polylines.push(line);
                 }
                 let annotates = match cursor.word()? {
-                    "D" => Some(ogeom_doc::Annotated::Dimension(cursor.count()? as usize)),
-                    "T" => Some(ogeom_doc::Annotated::Tolerance(cursor.count()? as usize)),
-                    "M" => Some(ogeom_doc::Annotated::Datum(cursor.count()? as usize)),
+                    "D" => Some(ogeom_doc::Annotated::Dimension(cursor.count()?)),
+                    "T" => Some(ogeom_doc::Annotated::Tolerance(cursor.count()?)),
+                    "M" => Some(ogeom_doc::Annotated::Datum(cursor.count()?)),
                     _ => None,
                 };
                 document.pmi_mut().callouts.push(ogeom_doc::Callout {
@@ -917,10 +917,10 @@ pub fn read_document(text: &str) -> OgeomResult<ogeom_doc::Document> {
                 } else {
                     None
                 };
-                let count = cursor.count()? as usize;
+                let count = cursor.count()?;
                 let mut callouts = Vec::with_capacity(count);
                 for _ in 0..count {
-                    callouts.push(cursor.count()? as usize);
+                    callouts.push(cursor.count()?);
                 }
                 document.add_view(ogeom_doc::View {
                     name,
@@ -933,7 +933,7 @@ pub fn read_document(text: &str) -> OgeomResult<ogeom_doc::Document> {
                 let author = read_text(&mut cursor)?;
                 let text_body = read_text(&mut cursor)?;
                 let product = if cursor.flag()? {
-                    let index = cursor.count()? as usize;
+                    let index = cursor.count()?;
                     ids.get(index).copied()
                 } else {
                     None
